Toivottavasti toimii

+++ PizzaGatto/src/fi/omapizzeria/pizzagatto/admin/LogInService.java
@@ -20,59 +20,62 @@ public class LogInService {
 	/*|*/												/*|*/
 	/*-----------------------------------------------------*/
 	
-	
-	
-	
 	/*
-	 * Kirjautuu sis‰‰n. Kirjautumisikkuna tulee WindowsService-luokasta ja
-	 * mahdollinen bannaus toteutetaan BanningService-luokan metodilla
+	 * Metodille l‰hetet‰‰n k‰ytt‰j‰n syˆtt‰m‰t tunnukset ja ip-osoite. Metodi palauttaa numeron, jonka perusteella k‰ytt‰j‰lle n‰ytet‰‰n haluttu sivu
 	 */
 	
 	public int tryLogin(String ipAddress, ArrayList<String> namepassword) throws ServletException, NoSuchAlgorithmException, UnsupportedEncodingException, DAOPoikkeus {
 		admindao = new AdminDAO();
 		ArrayList <String> legit = new ArrayList<String>(admindao.getAdmin());
-		System.out.println(legit.get(0) + "  " + legit.get(1));
-		AdminUser ad = new AdminUser(namepassword.get(0), namepassword.get(1));
-		
-		
-		boolean dontAdd = false; //Kello 3:40 purkkaviritelm‰n huipentuma
+		AdminUser ad = new AdminUser(namepassword.get(0), namepassword.get(1), legit.get(2));
+		ad.setPasswordHash(legit.get(1));
+		boolean dontAdd = false;
 
+		/*
+		 * K‰yd‰‰n l‰pi kirjautumisyritt‰jien ip-osoitteet. Jos k‰ytt‰j‰n osoitete lˆytyy jo listalta sit‰ ei lis‰t‰.
+		 */
 		
-		for (int i = 0; i < banService.getAttempterIps().size(); i++) {		//Horrible shit coming...
+		for (int i = 0; i < banService.getAttempterIps().size(); i++) {
 			if (banService.getAttempterIps().get(i).getIp().equals(ipAddress)) {	
 				dontAdd = true;
 			}
 		}
 		
-		if (dontAdd == false) {		//Ohjelmointikurssista 0/5
+		if (dontAdd == false) {
 			banService.addAttempter(ipAddress);
 		}
 		
 		/*
-		 * T‰m‰ puolestaan on reiki‰ t‰ynn‰. Kai.
+		 * Verrataan tietokannassa olevaa k‰ytt‰j‰tunnusta syˆtettyyn k‰ytt‰j‰tunnukseen ja
+		 * tietokannassa olevaa salattua salasanaa k‰ytt‰j‰n syˆtt‰m‰‰n salasanaan, joka on salattu
+		 * k‰ytt‰en samaa suolaa kuin adminin salasanassa
 		 */
-		if (ad.vertaaSalasanaa(namepassword.get(1))) {
+		
+		if (ad.vertaaSalasanaa(namepassword.get(1)) && namepassword.get(0).equals(legit.get(0))) {
 			banService.removeAttempter(ipAddress);
+			//Arvolla 1 k‰ytt‰j‰ p‰‰see sis‰‰n
 			return 1;
-		}
+		} 
 		for (int i = 0; i < banService.getAttempterIps().size(); i++) {
-			 if (banService.getAttempterIps().get(i).getIp().equals(ipAddress)) {
-				
+			// K‰ytt‰j‰ on t‰ss‰ vaiheessa antanut v‰‰r‰t tunnukset
+			if (banService.getAttempterIps().get(i).getIp().equals(ipAddress)) {
+				//Kasvatetaan kirjautumiskertoja
 				try {
 					admindao = new AdminDAO();
 					admindao.riseAttempterTries(ipAddress);
 				} catch (DAOPoikkeus e) {
 					throw new ServletException(e);
 				}
+				//Jos kirjautumiskertoja tulee liikaa k‰ytt‰j‰ b‰nn‰t‰‰n
 				if (banService.getAttempterIps().get(i).getTries() >= maxTries) {
 					banService.banUser(ipAddress);
 					banService.removeAttempter(ipAddress);
+					//Arvolla 0 k‰ytt‰j‰ b‰nn‰t‰‰n
 					return 0;
 				} 
 			}
 		}
+		//Arvolla 2 k‰ytt‰j‰lle kerrotaan ett‰ annetut tunnukset ovat v‰‰r‰t
 		return 2;			
 	}
 }
-
-//namepassword.get(0).equals(legitId) && namepassword.get(1).equals(legitPassword)
